gobot/services/routing-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
ortools>=9.8.0
